requirements.txt
Django==4.2.0
django-mongoengine==0.5.6
djangorestframework==3.14.0
django-cors-headers==4.1.0
django-allauth==0.54.0
Pillow==9.5.0
python-dotenv==1.0.0
django-rest-auth==0.9.5
